* 'master' of https://github.com/premnath1989/TMLI:
  Rp
  IDR
  back
  Lain-lain and listing name
  add alamat cantor
  Si Menu, back button, hubungan lain2
  Prospect - Existing List.
  hot fix folder
  foldering in product information
  product information folding fix
  Prospect.
  Prospect.
  e-library add http mode and folder function
  Prospect
  Prospect
  Prospect
  - Add Webservice for Product Information  - 120 hari data deletion
  Prospect

* 'master' of https://github.com/premnath1989/TMLI:
  Rp
  IDR
  back
  Lain-lain and listing name
  add alamat cantor
  Si Menu, back button, hubungan lain2
  Prospect - Existing List.
  hot fix folder
  foldering in product information
  product information folding fix
  Prospect.
  Prospect.
  e-library add http mode and folder function
  Prospect
  Prospect
  Prospect
  - Add Webservice for Product Information  - 120 hari data deletion
  Prospect

# Conflicts:
#	MobileOfficeSolution/MobileOfficeSolution/Info.plist
#	MobileOfficeSolution/MobileOfficeSolution/String.h
#	MobileOfficeSolution/MobileOfficeSolution/String.m
#	MobileOfficeSolution/MobileOfficeSolution/database/query/MOSDB.sqlite.sql

+++ MobileOfficeSolution/MobileOfficeSolution/TableManagement.h
@@ -17,7 +17,7 @@
 
 - (instancetype)init:(UIView *)view themeColour:(UIColor *)Colour themeFont:(UIFont *)font;
 - (UIView *) TableHeaderSetup:(NSArray *)columnHeaders positionY:(float)originY;
-- (void)TableRowInsert:(NSMutableArray *)dataArray index:(NSInteger)index table:(UITableViewCell*)cell color:(UIColor *)textColor;
+- (void)TableRowInsert:(NSMutableArray *)dataArray index:(int)index table:(UITableViewCell*)cell color:(UIColor *)textColor;
 - (UIView *) TableHeaderSetupXY:(NSArray *)columnHeaders positionY:(float)originY positionX:(float)originX;
 
 @end
